rewrite Search and Sidebar Components
