--- src/components/Search.module.css
@@ -1,14 +1,0 @@
-.content {
-    height: calc(100vh - 40px);
-    overflow-y: scroll;
-}
-
-.row {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-}
-
-.item {
-    margin: 20px;
-}
